+++ tabbycat/adjallocation/templates/EditPanelAdjudicatorsContainer.vue
@@ -1,40 +1,75 @@
 <template>
-
-  <drag-and-drop-layout :unallocatedItems="unallocatedItems"
-                        :unallocatedComponent="unallocatedComponent"
-                        :handle-unused-drop="moveAdjudicator">
-
-    <drag-and-drop-actions slot="actions" :count="debatesOrPanelsCount" prioritise="true" allocate="true" shard="true"
-                           @show-shard="showShard" @show-allocate="showAllocate" @show-prioritise="showPrioritise">
+  <drag-and-drop-layout
+    :unallocatedItems="unallocatedItems"
+    :unallocatedComponent="unallocatedComponent"
+    :handle-unused-drop="moveAdjudicator"
+  >
+    <drag-and-drop-actions
+      slot="actions"
+      :count="debatesOrPanelsCount"
+      prioritise="true"
+      allocate="true"
+      shard="true"
+      @show-shard="showShard"
+      @show-allocate="showAllocate"
+      @show-prioritise="showPrioritise"
+    >
       <template slot="extra-actions">
-        <button :class="['btn', debatesOrPanelsCount > 0 ? 'btn-outline-primary' : 'btn-success']"
-                @click="showCreatePanels" v-text="gettext('Create Panels')"></button>
+        <button
+          :class="['btn', debatesOrPanelsCount > 0 ? 'btn-outline-primary' : 'btn-success']"
+          @click="showCreatePanels"
+          v-text="gettext('Create Panels')"
+        ></button>
       </template>
       <template slot="default-highlights">
-        <button class="btn conflictable conflicts-toolbar hover-histories-2-ago"
-                data-toggle="tooltip" v-text="gettext('Seen')"
-                title="This adjudicator has judged with this adjudicator previously"></button>
-        <button class="btn conflictable conflicts-toolbar hover-institution"
-                data-toggle="tooltip" v-text="gettext('Institution')"
-                title="This adjudicator is from the same institution as this panelist."></button>
-        <button class="btn conflictable conflicts-toolbar hover-adjudicator"
-                data-toggle="tooltip" v-text="gettext('Conflict')"
-                title="This adjudicator has a nominated conflict with this panelist."></button>
-        <button class="btn panel-incomplete"
-                data-toggle="tooltip" v-text="gettext('Missing')"
-                title="Panel is either missing a chair or enough adjudicators for a voting majority."></button>
+        <button
+          class="btn conflictable conflicts-toolbar hover-histories-2-ago"
+          data-toggle="tooltip"
+          v-text="gettext('Seen')"
+          title="This adjudicator has judged with this adjudicator previously"
+        ></button>
+        <button
+          class="btn conflictable conflicts-toolbar hover-institution"
+          data-toggle="tooltip"
+          v-text="gettext('Institution')"
+          title="This adjudicator is from the same institution as this panelist."
+        ></button>
+        <button
+          class="btn conflictable conflicts-toolbar hover-adjudicator"
+          data-toggle="tooltip"
+          v-text="gettext('Conflict')"
+          title="This adjudicator has a nominated conflict with this panelist."
+        ></button>
+        <button
+          class="btn panel-incomplete"
+          data-toggle="tooltip"
+          v-text="gettext('Missing')"
+          title="Panel is either missing a chair or enough adjudicators for a voting majority."
+        ></button>
       </template>
     </drag-and-drop-actions>
 
     <template slot="debates">
-      <drag-and-drop-debate v-for="panel in sortedDebatesOrPanels" :key="panel.pk" :debateOrPanel="panel">
-        <debate-or-panel-importance slot="importance"
-                                    :debate-or-panel="panel"></debate-or-panel-importance>
-        <debate-or-panel-adjudicators slot="adjudicators" :debate-or-panel="panel"
-                                      :handle-debate-or-panel-drop="moveAdjudicator">
+      <drag-and-drop-debate
+        v-for="panel in sortedDebatesOrPanels"
+        :key="panel.pk"
+        :debateOrPanel="panel"
+      >
+        <debate-or-panel-importance
+          slot="importance"
+          :debate-or-panel="panel"
+        ></debate-or-panel-importance>
+        <debate-or-panel-adjudicators
+          slot="adjudicators"
+          :debate-or-panel="panel"
+          :handle-debate-or-panel-drop="moveAdjudicator"
+          :handle-panel-swap="swapPanels"
+        >
         </debate-or-panel-adjudicators>
-        <template slot="teams"><span></span></template><!--Hide Teams-->
-        <template slot="venue"><span></span></template><!--Hide Venues-->
+        <template slot="teams"><span></span></template
+        ><!--Hide Teams-->
+        <template slot="venue"><span></span></template
+        ><!--Hide Venues-->
       </drag-and-drop-debate>
       <div class="text-center lead mx-5 p-5" v-if="debatesOrPanelsCount === 0">
         <p class="mx-5 lead mt-2 px-5" v-text="gettext(createPanelsInline)"></p>
@@ -45,14 +80,17 @@
       <modal-for-creating-preformed-panels :context-of-action="'create_preformed_panels'">
       </modal-for-creating-preformed-panels>
       <modal-for-sharding :intro-text="gettext(shardIntro)"></modal-for-sharding>
-      <modal-for-allocating :intro-text="gettext(allocateIntro)" :for-panels="true"
-                            :context-of-action="'allocate_panel_adjs'"></modal-for-allocating>
-      <modal-for-prioritising :intro-text="gettext(prioritiseIntro)"
-                              :context-of-action="'prioritise_panels'"></modal-for-prioritising>
+      <modal-for-allocating
+        :intro-text="gettext(allocateIntro)"
+        :for-panels="true"
+        :context-of-action="'allocate_panel_adjs'"
+      ></modal-for-allocating>
+      <modal-for-prioritising
+        :intro-text="gettext(prioritiseIntro)"
+        :context-of-action="'prioritise_panels'"
+      ></modal-for-prioritising>
     </template>
-
   </drag-and-drop-layout>
-
 </template>
 
 <script>